implementedStrategyDesignPatternForSort
