+++ Datastructure&Algorithmen/Algorithmen/Algorithmen.csproj
@@ -1,14 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Library</OutputType>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Algorithmen\Algorithmen.csproj" />
     <ProjectReference Include="..\Common\Common.csproj" />
   </ItemGroup>
