+++ Datastructure&Algorithmen/Common/ISortStrategy.cs
@@ -1,0 +1,7 @@
+﻿namespace Common
+{
+    public interface ISortStrategy<T>
+    {
+        public void Sort(Node<T> head);
+    }
+}
