+++ Datastructure&Algorithmen/DatastructureTests/SingleLinkedListTests.cs
@@ -185,7 +185,7 @@ namespace DatastructureTests
             linkedList.AddLast(_Person2);
             linkedList.AddLast(_Person3);
             linkedList.AddLast(_PersonNull);
-            linkedList.BubbleSort();
+            linkedList.Sort();
             Assert.That(linkedList.GetAllNodesData(), Is.EqualTo(new List<Person> { _PersonNull, _Person2, _Person1, _Person3 } ));
         }
     }
