+++ Datastructure&Algorithmen/Datastructure_and_Algorithmen.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36429.23
+# Visual Studio Version 18
+VisualStudioVersion = 18.0.11123.170 d18.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Datastructure", "Datastructure\Datastructure.csproj", "{B576225F-62BE-46FD-BB1B-38279223BD1A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{029C82AB-5849-4181-8B09-65DC22D89A47}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DatastructureTests", "DatastructureTests\DatastructureTests.csproj", "{590AA66B-D950-42BE-8683-8291A670888B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Algorithmen", "Algorithmen\Algorithmen.csproj", "{D45EB65E-DC45-4D73-A4A9-4CFC4BB2F9C7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{590AA66B-D950-42BE-8683-8291A670888B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{590AA66B-D950-42BE-8683-8291A670888B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{590AA66B-D950-42BE-8683-8291A670888B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D45EB65E-DC45-4D73-A4A9-4CFC4BB2F9C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D45EB65E-DC45-4D73-A4A9-4CFC4BB2F9C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D45EB65E-DC45-4D73-A4A9-4CFC4BB2F9C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D45EB65E-DC45-4D73-A4A9-4CFC4BB2F9C7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
